Breaking CFJsonItem part 2: extracting datacube extension code
